better build support for warnings/errors/fatals

+++ Engine/Editor/Source/UI/Tools/AssetEditors/TextureEditor/TextureEditorTool.cpp
@@ -461,7 +461,7 @@ namespace Lumina
                         }
                         else
                         {
-                            ImGui::Text("%llu B", actualSize);
+                            ImGui::Text("%llu B", (unsigned long long)actualSize);
                         }
         
                         // Percentage
